refactor(store): break array-bookkeeping <-> create-form-store cycle
array-bookkeeping needs the FieldRecord / ElementRecord / OriginalsRecord
shapes, which create-form-store defined, while create-form-store imports
createArrayBookkeeping back from array-bookkeeping. The back-edge is
type-only (erased at compile, so no runtime cycle), but fallow flags it
and it muddies the module graph.

Extract the three record types into a new store-records.ts leaf (they
depend only on Path). create-form-store and array-bookkeeping both import
from the leaf; the bookkeeping module no longer imports the store. The
helper functions over FieldRecord stay in create-form-store. tsc --noEmit
and eslint pass; the cycle is gone.
